requirements.txt
coqui-tts>=0.25.0
transformers>=4.30.0,<4.46.0
torch>=2.0.0
torchaudio>=2.0.0
numpy
scipy
pydub
pyloudnorm
nltk
num2words
tqdm
pytest
# ── Video pipeline dependencies ───────────────────────────────────────────
whisperx
pyannote.audio>=3.1.1
Pillow>=10.0.0
opencv-python
ffmpeg-python
